+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 2.3.0
 -----
 
- * added JsonResponse::prefixJson() to mitigate CRSF attacks by prepending a prefix
+ * added JsonResponse::setPrefix() to mitigate CRSF attacks by prepending a prefix
    to the generated JSON
 
 2.2.0