+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+2.3.0
+-----
+
+ * added JsonResponse::setPrefix() to mitigate CRSF attacks by prepending a prefix
+   to the generated JSON
+
 2.2.0
 -----